requirements.txt
flask>=3.1.2
python-dotenv>=1.1.1
flask-sqlalchemy>=3.1.1
flask-migrate>=4.1.0
psycopg2-binary>=2.9.10
asyncpg>=0.30.0
pika>=1.3.2